Update max_allowed_packet for cliennt connections.
MySQL client also has a configuration value for `max_allowed_packet` and this can be different from the server (which in our case is 64M). This is currently using the default value of 16M. This causes issues when performing MySQL operations directly against the DB (PHP relies on the server configuration and is not subject to this problem), specifically mysqldump (`drush sql-dump`). If a row is inserted that is higher than 16M the mysql client cannot retrieve the value.